Log out

+++ src/components/Post.jsx
@@ -1,6 +1,7 @@
 import CardBase from "./CardBase";
 import Image from "next/image";
 import Button from "./Button";
+import { toast } from "react-hot-toast";
 import { useRouter } from "next/router";
 import {
   ChatBubbleLeftIcon,
@@ -20,16 +21,38 @@ export default function Post({
 }) {
   const router = useRouter();
   const handleClose = () => {
-    onDelete(postID);
+    const confirmDelete = confirm("¿Estas seguro que quieres borrar el post?");
+    if (confirmDelete) {
+      onDelete(postID);
+    }
     return;
   };
-  const handleEdit = () => {
-    localStorage.setItem(
-      "postInfo",
-      JSON.stringify({ title: mainText, body: body, id: postID }),
-    );
-    router.push("http://localhost:3000/new?state=edit");
-    return;
+  const handleEdit = async () => {
+    try {
+      const userLogged = JSON.parse(localStorage.getItem("currentUser"))._id;
+      const values = { postID: postID, userLogged: userLogged };
+      const endpoint = "https://apidevto.onrender.com/auth/verifyUser";
+      const response = await fetch(endpoint, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(values),
+      });
+      const data = await response.json();
+
+      if (data.success) {
+        localStorage.setItem(
+          "postInfo",
+          JSON.stringify({ title: mainText, body: body, id: postID }),
+        );
+        router.push("/new?state=edit");
+        return;
+      } else {
+        toast.error("Usuario sin permiso de edicion");
+      }
+    } catch (error) {
+      console.error("Hubo un error al hacer la peticion", error);
+      toast.error("Error con el servidor. Intente mas tarde");
+    }
   };
   return (
     <CardBase className="md:p-4">
@@ -37,6 +60,7 @@ export default function Post({
         <div className="flex items-center gap-2">
           <Image
             src={userPicture}
+            alt="userPicture"
             width={30}
             height={30}
             className="h-8 w-8 rounded-full"
